refactor: Extract BulkCreateEpisodesButton
